+++ src/python/bisect/index.ts
@@ -1,0 +1,3 @@
+export { bisect } from './bisect.ts'
+export { bisect_left } from './bisect_left.ts'
+export { bisect_right } from './bisect_right.ts'